move environment to its own package in core (#21)
