+++ core/src/instructions/mod.rs
@@ -1,6 +1,8 @@
 pub mod error;
 pub mod val;
 use crate::constraint::Constraint;
+use crate::environment::EnvExtension;
+use crate::environment::EnvExtensionRecord;
 use crate::memory::*;
 use crate::stack::*;
 use error::InstructionError;
@@ -64,22 +66,6 @@ where
             constraints: None,
         }
     }
-}
-
-pub trait EnvExtensionRecord: Sized {
-    fn apply<E: EnvExtension>(&self, env: E) -> Result<E, E::ErrorType>;
-}
-
-pub trait EnvExtension: Clone {
-    type InnerValue;
-    type ErrorType: std::fmt::Debug;
-    type IndexType;
-    type DiffRecordType: EnvExtensionRecord;
-
-    fn write<V: Into<Self::InnerValue>>(&self, v: V) -> Result<Self, Self::ErrorType>
-    where
-        Self: Sized;
-    fn read<I: Into<Self::IndexType>>(&self, idx: I) -> Result<Self::InnerValue, Self::ErrorType>;
 }
 
 pub trait AbstractInstruction<S, M, Extension, StepResult>
